Add bandit as a development dependency
Add bandit as a development dependency and configure it to allow the use
of `assert` in pytest files.

Add a bandit hook to run the `bandit` command with the
`--configfile=pyproject.toml` flag to lint Python files.

Configure VS Code to use bandit with the `--configfile=pyproject.toml`
flag for linting.

Add a badge for mypy to the `README.md` file.

See also
* https://bandit.readthedocs.io/en/latest/plugins/b101_assert_used.html#b101-test-for-use-of-assert,
* https://bandit.readthedocs.io/en/latest/config.html#bandit-settings, and
* https://code.visualstudio.com/docs/python/settings-reference#_bandit.